Refactor: Streamline Vehicle & Rover Models
