+++ src/main/java/marsmission/marsrover/model/Vehicle.java
@@ -1,36 +1,37 @@
 package marsmission.marsrover.model;
 
-import java.util.Arrays;
-
 public abstract class Vehicle {
-    private static int existingInstances = 0;
-    final int ID;
+    final long ID;
+    final String CODENAME;
     final VehicleType TYPE;
     char direction;
     int[] position;
 
     public Vehicle(VehicleType type, char direction, int[] position) {
-        this.ID = generateId();
+        Identifier identifier = new Identifier();
+        this.ID = identifier.generateId();
+        this.CODENAME = generateCodeName();
         this.TYPE = type;
         this.direction = direction;
         this.position = position;
     }
 
-    private int generateId() {
-        int newId = existingInstances;
-        existingInstances++;
-
-        return newId;
+    private String generateCodeName() {
+        return getType().toString() + getId();
     }
 
-    public void getDetails() {
-        System.out.println("Type: " + TYPE.toString());
+    public void displayInfo() {
         System.out.println("ID: " + ID);
-        System.out.println("Start Position: " + direction + " " + Arrays.toString(position));
+        System.out.println("Codename: " + CODENAME);
+        System.out.println("Type: " + TYPE.toString());
     }
 
-    public int getId() {
+    public long getId() {
         return ID;
+    }
+
+    public String getCodeName() {
+        return CODENAME;
     }
 
     public VehicleType getType() {
